プレイ共通処理クラス化、リザルト移行処理
