+++ web_ts/src/Game.ts
@@ -3,6 +3,7 @@ import Preloader from './Preloader';
 import MainMenu from './MainMenu';
 import Omote from './Omote';
 import Ura from './Ura';
+import Result from './Result';
 
 export default class Game extends Phaser.Game {
 
@@ -14,6 +15,8 @@ export default class Game extends Phaser.Game {
 		this.state.add('MainMenu', MainMenu, false);
 		this.state.add('Omote', Omote, false);
 		this.state.add('Ura', Ura, false);
+		this.state.add('Result', Result, false);
+
 
 		this.state.start('Boot');
 	}
